requirements.txt
beeprint
javascript
hexbytes
numpy
pandas
pymongo
python-dotenv
requests
setuptools
web3>=6.0.0
